+++ src/event/equip_panel_run_equipment_list_thread.c
@@ -143,7 +143,7 @@ void equip_panel_run_equipment_list_thread(void) {
         equip_gfx_build_scaled_draw_area_packets(&screen->portrait, &g_equip_character_status_frame_rect[2], frame,
             g_main_gfx_screen_polarity * 0xf0, (const s16*)thread);
         equip_panel_set_primitive_colors(screen, thread);
-        equip_gfx_apply_menu_palette_for_mode((s32)&screen->numeric_frame, (s32*)thread);
+        equip_gfx_apply_menu_palette_for_mode(&screen->numeric_frame, (s32*)thread);
         battle_gfx_draw_or_append_gpu_primitive((s32*)&screen->portrait.areas[1]);
         battle_gfx_draw_or_append_gpu_primitive((s32*)&screen->draw_offsets[1]);
         battle_gfx_draw_or_append_gpu_primitive((s32*)&screen->sprites[0]);
